refactor: move queueing code into its own directory
