+++ tests/queues/fixed-priority-queue.test.mjs
@@ -1,7 +1,7 @@
 import { expect } from 'expect'
 
-import { FixedPriorityQueue } from '../lib/fixed-priority-queue.cjs'
-import { defaultQueueSize } from '../lib/utility-types.cjs'
+import { FixedPriorityQueue } from '../../lib/queues/fixed-priority-queue.cjs'
+import { defaultQueueSize } from '../../lib/queues/queue-types.cjs'
 
 describe('Fixed priority queue test suite', () => {
   it('Verify constructor() behavior', () => {
